cast to double.

+++ google_benchmark/art_insert_lookup_benchmark.cpp
@@ -118,7 +118,7 @@ BENCHMARK_DEFINE_F(INT32_ART_Fixture, SortedDenseKeys_Insert_Test)(benchmark::St
 			state.ResumeTiming();
 		}
         Insert();
-        state.counters["Throughput"] = benchmark::Counter(state.range(0), benchmark::Counter::kIsRate);
+        state.counters["Throughput"] = benchmark::Counter(static_cast<double>(state.range(0)), benchmark::Counter::kIsRate);
     }
 }
 
@@ -133,7 +133,7 @@ BENCHMARK_DEFINE_F(INT32_ART_Fixture, SortedDenseKeys_Lookup_Test)(benchmark::St
             state.ResumeTiming();
         }
         Lookup();
-        state.counters["Throughput"] = benchmark::Counter(state.range(0), benchmark::Counter::kIsRate);
+        state.counters["Throughput"] = benchmark::Counter(static_cast<double>(state.range(0)), benchmark::Counter::kIsRate);
     }
 }
 BENCHMARK_REGISTER_F(INT32_ART_Fixture, SortedDenseKeys_Lookup_Test)->Arg(1000)->Arg(1000000)->Unit(benchmark::kMillisecond);
@@ -147,7 +147,7 @@ BENCHMARK_DEFINE_F(INT32_ART_Fixture, RandomDenseKeys_Insert_Test)(benchmark::St
             state.ResumeTiming();
         }
         Insert();
-        state.counters["Throughput"] = benchmark::Counter(state.range(0), benchmark::Counter::kIsRate);
+        state.counters["Throughput"] = benchmark::Counter(static_cast<double>(state.range(0)), benchmark::Counter::kIsRate);
     }
 }
 
@@ -162,7 +162,7 @@ BENCHMARK_DEFINE_F(INT32_ART_Fixture, RandomDenseKeys_Lookup_Test)(benchmark::St
             state.ResumeTiming();
         }
         Lookup();
-        state.counters["Throughput"] = benchmark::Counter(state.range(0), benchmark::Counter::kIsRate);
+        state.counters["Throughput"] = benchmark::Counter(static_cast<double>(state.range(0)), benchmark::Counter::kIsRate);
     }
 }
 BENCHMARK_REGISTER_F(INT32_ART_Fixture, RandomDenseKeys_Lookup_Test)->Arg(1000)->Arg(1000000)->Unit(benchmark::kMillisecond);
@@ -175,7 +175,7 @@ BENCHMARK_DEFINE_F(INT32_ART_Fixture, SparseUniqueKeys_Insert_Test)(benchmark::S
             state.ResumeTiming();
         }
         Insert();
-        state.counters["Throughput"] = benchmark::Counter(state.range(0), benchmark::Counter::kIsRate);
+        state.counters["Throughput"] = benchmark::Counter(static_cast<double>(state.range(0)), benchmark::Counter::kIsRate);
     }
 }
 
@@ -190,7 +190,7 @@ BENCHMARK_DEFINE_F(INT32_ART_Fixture, SparseUniqueKeys_Lookup_Test)(benchmark::S
             state.ResumeTiming();
         }
         Lookup();
-        state.counters["Throughput"] = benchmark::Counter(state.range(0), benchmark::Counter::kIsRate);
+        state.counters["Throughput"] = benchmark::Counter(static_cast<double>(state.range(0)), benchmark::Counter::kIsRate);
     }
 }
 BENCHMARK_REGISTER_F(INT32_ART_Fixture, SparseUniqueKeys_Lookup_Test)->Arg(1000)->Arg(1000000)->Unit(benchmark::kMillisecond);
